watchtower: make use of the new tower-to-session index
In this commit, the towerID-to-sessionID index added in the previous
commit is put to use in order to make session lookup more efficient in
certain places. In the process, 2 TODO's are also removed from the code.

(cherry picked from commit ecd2eb965a75fd1a0dc55fe533f5ac8ebe529209)